scraping website

+++ test.py
@@ -16,19 +16,14 @@ with open("FOLDER/info.dat") as infoFile:
     if info["_version"] != "2.0.0":
         print("beatmap not version 2.0.0")
     folder += info["_songName"]
+if(folder == "Songs/"):
+    quit()
 if os.path.exists(folder):
-    command = f"rm -rf {folder}"
-    if(goofy):
-        command = f"powershell -command '{command}'"
-    os.system(command)
+    shutil.rmtree(folder)
 command = f"mv FOLDER {folder}"
 if(goofy):
     command = f"powershell -command '{command}'"
 os.system(command)
 os.remove(filename)
-shutil.rmtree("FOLDER")
-"""if os.path.exists("FOLDER"):
-    command = "rm -rf FOLDER"
-    if(goofy):
-        command = f"powershell -command '{command}'"
-    os.system(command)"""
+if os.path.exists("FOLDER"):
+    shutil.rmtree("FOLDER")
